// frontend/src/app/api/model/providerSetupInitResponse.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { SetupDimensionDto } from './setupDimensionDto';


export interface ProviderSetupInitResponse { 
    canSaveToServer?: boolean;
    secretsYmlPath?: string;
    quercheckerYmlPath?: string;
    dimensions?: Array<SetupDimensionDto>;
}