// frontend/src/app/api/model/providerSetupSaveRequest.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ProviderSetupSaveRequest { 
    secrets?: { [key: string]: string; };
    config?: { [key: string]: string; };
    searchProvider?: string;
    llmProvider?: string;
}